requirements.txt
Pillow==8.0.1
selenium==3.141.0
tqdm==4.60.0
urllib3==1.26.3
imagehash==4.2.1